+++ modules/components/src/test/java/app/popups/PopupTest.java
@@ -1,129 +1,59 @@
 package app.popups;
 
 import io.github.palexdev.mfxcomponents.controls.buttons.MFXButton;
+import io.github.palexdev.mfxcomponents.controls.buttons.MFXFilledButton;
+import io.github.palexdev.mfxcomponents.controls.buttons.MFXTextButton;
 import io.github.palexdev.mfxcomponents.theming.enums.MFXThemeManager;
-import io.github.palexdev.mfxcomponents.window.popups.MFXPopup;
+import io.github.palexdev.mfxcomponents.window.MFXRichContent;
 import io.github.palexdev.mfxcomponents.window.popups.MFXTooltip;
-import io.github.palexdev.mfxcore.base.beans.Position;
-import io.github.palexdev.mfxcore.base.beans.Size;
 import io.github.palexdev.mfxcore.builders.InsetsBuilder;
-import io.github.palexdev.mfxcore.utils.PositionUtils;
-import io.github.palexdev.mfxcore.utils.fx.CSSFragment;
 import javafx.application.Application;
 import javafx.collections.FXCollections;
-import javafx.geometry.Point2D;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
 import javafx.scene.control.ComboBox;
-import javafx.scene.control.Label;
-import javafx.scene.input.MouseButton;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
-import org.scenicview.ScenicView;
 
 public class PopupTest extends Application {
+    private final String lorem = "Lorem ipsum dolor sit amet. " +
+        "Aut nulla inventore in tempore suscipit ab possimus soluta id rerum maiores. " +
+        "Vel voluptatem unde ea reprehenderit labore ut minima ratione. " +
+        "Ut quos quam in tempora quibusdam id iste accusamus aut dolores facere et Quis dolore.";
 
     @Override
     public void start(Stage primaryStage) {
         StackPane pane = new StackPane();
 
-        MFXButton button = MFXButton.filled();
-        button.setText("Show Popup");
+        ComboBox<Pos> anchors = new ComboBox<>(FXCollections.observableArrayList(Pos.values()));
+        anchors.getItems().removeAll(Pos.BASELINE_CENTER, Pos.BASELINE_LEFT, Pos.BASELINE_RIGHT);
+        anchors.getSelectionModel().select(Pos.BOTTOM_CENTER);
+        StackPane.setAlignment(anchors, Pos.TOP_CENTER);
+        StackPane.setMargin(anchors, InsetsBuilder.top(20));
 
-        ComboBox<Pos> positions = new ComboBox<>(FXCollections.observableArrayList(Pos.values()));
-        positions.getItems().removeAll(Pos.BASELINE_CENTER, Pos.BASELINE_LEFT, Pos.BASELINE_RIGHT);
-        positions.getSelectionModel().select(Pos.BOTTOM_CENTER);
-        positions.setEditable(false);
-        StackPane.setAlignment(positions, Pos.TOP_CENTER);
-        StackPane.setMargin(positions, InsetsBuilder.top(80));
-
-/*        MFXButton button2 = MFXButton.filled();
-        button2.setText("Show Popup2");
-        StackPane.setMargin(button2, InsetsBuilder.left(400));
-
-        MFXTooltip popup = new MFXTooltip(button);
-        popup.setContent(new SimpleContent());
-        popup.install();
-
-        MFXTooltip popup2 = new MFXTooltip(button2);
-        popup2.setContent(new SimpleContent("2nd Tooltip"));
-        popup2.install();*/
+        MFXButton button = new MFXFilledButton("Show Popup");
 
         MFXTooltip tooltip = new MFXTooltip(button);
-        tooltip.setContent(new SimpleContent());
-        button.setMFXTooltip(tooltip);
-
-        MFXPopup popup = new MFXPopup();
-        popup.setContent(new SimpleContent());
-        button.addEventHandler(MouseEvent.MOUSE_CLICKED, e -> {
-            if (e.getButton() == MouseButton.PRIMARY) {
-                popup.show(button, positions.getValue());
-            } else {
-                popup.show(primaryStage, positions.getValue());
-            }
+        tooltip.setContent(new MFXRichContent());
+        tooltip.setHeader("A tooltip duh!");
+        tooltip.setText(lorem);
+        tooltip.setPrimaryAction(() -> {
+            MFXButton primary = new MFXTextButton("Learn More");
+            primary.setOnAction(e -> tooltip.setText("I already feel smarter"));
+            return primary;
         });
-
-        Runnable updateOffset = () -> {
-            Pos pos = positions.getValue();
-            if (pos == null || pos == Pos.CENTER) {
-                popup.setOffset(Position.origin());
-                return;
-            }
-
-            double x = 0;
-            double y = 0;
-            if (PositionUtils.isLeft(pos) || PositionUtils.isRight(pos)) x = 8;
-            if (PositionUtils.isTop(pos) || PositionUtils.isBottom(pos)) y = 8;
-            popup.setOffset(Position.of(x, y));
-        };
-        positions.valueProperty().addListener(i -> updateOffset.run());
-        updateOffset.run();
-
-        Label screen = new Label();
-        pane.addEventFilter(MouseEvent.MOUSE_PRESSED, e -> {
-            if (e.getButton() == MouseButton.SECONDARY) {
-                Point2D toScreen = button.localToScreen(0, 0);
-                screen.setText(Position.of(toScreen.getX(), toScreen.getY()).toString());
-            }
+        tooltip.setSecondaryAction(() -> {
+            MFXButton secondary = new MFXTextButton("Cancel");
+            secondary.setOnAction(e -> tooltip.setText(lorem));
+            return secondary;
         });
-        StackPane.setAlignment(screen, Pos.TOP_CENTER);
-        StackPane.setMargin(screen, InsetsBuilder.top(12));
+        tooltip.install(button);
+        anchors.valueProperty().addListener((ob, o, n) -> tooltip.setAnchor(n));
 
-        Label content = new Label();
-        content.textProperty().bind(popup.contentBoundsProperty()
-            .map(b -> Size.of(b.getWidth(), b.getHeight()).toString()));
-        StackPane.setAlignment(content, Pos.TOP_CENTER);
-        StackPane.setMargin(content, InsetsBuilder.top(42));
-
-        CSSFragment.Builder.build()
-            .addSelector(".mfx-popup .content")
-            .addSelector(".mfx-tooltip .content")
-            .addStyle("-fx-background-color: rgba(0, 0, 0, 0.2)")
-            .addStyle("-fx-background-radius: 12px")
-            .addStyle("-fx-padding: 12px")
-            //.addStyle("-mfx-offset: \"0 8\"")
-            .closeSelector()
-            .applyOn(pane);
-
-        pane.getChildren().addAll(positions, button, screen, content);
+        pane.getChildren().addAll(button, anchors);
         Scene scene = new Scene(pane, 600, 600);
         MFXThemeManager.PURPLE_LIGHT.addOn(scene);
         primaryStage.setScene(scene);
         primaryStage.show();
-
-        ScenicView.show(scene);
-    }
-
-    private static class SimpleContent extends StackPane {
-        public SimpleContent() {
-            Label l = new Label("This is a simple popup with just a Label!");
-            getChildren().add(l);
-        }
-
-        public SimpleContent(String text) {
-            Label l = new Label(text);
-            getChildren().add(l);
-        }
     }
 }